Refactor llm Docsum (#1101)
