+++ comps/llms/src/doc-summarization/Dockerfile
@@ -19,10 +19,10 @@ COPY comps /home/user/comps
 
 RUN pip install --no-cache-dir --upgrade pip setuptools && \
     if [ ${ARCH} = "cpu" ]; then pip install --no-cache-dir torch torchvision --index-url https://download.pytorch.org/whl/cpu; fi && \
-    pip install --no-cache-dir -r /home/user/comps/llms/summarization/vllm/langchain/requirements.txt
+    pip install --no-cache-dir -r /home/user/comps/llms/src/doc-summarization/requirements.txt
 
 ENV PYTHONPATH=$PYTHONPATH:/home/user
 
-WORKDIR /home/user/comps/llms/summarization/vllm/langchain
+WORKDIR /home/user/comps/llms/src/doc-summarization
 
 ENTRYPOINT ["bash", "entrypoint.sh"]
