--- comps/llms/summarization/vllm/langchain/requirements-runtime.txt
@@ -1,1 +1,0 @@
-langserve
