--- comps/llms/summarization/vllm/langchain/requirements.txt
@@ -1,16 +1,0 @@
-docarray[full]
-fastapi
-httpx==0.27.2
-huggingface_hub
-langchain #==0.1.12
-langchain-huggingface
-langchain-openai
-langchain_community
-langchainhub
-opentelemetry-api
-opentelemetry-exporter-otlp
-opentelemetry-sdk
-prometheus-fastapi-instrumentator
-shortuuid
-transformers
-uvicorn
